Add image rendering
This adds image rendering using the `image` crate, which currently
supports both `jpeg` and `png` images.

Images are always fitted to the screen, based on which dimension of the
image is closest to the window dimensions. For the opposite dimension a
focus point can be set to allow control over how the additional content
is distributed.